+++ Source/AliveLibAE/ThrowableArray.cpp
@@ -150,7 +150,7 @@ BaseGameObject* ThrowableArray::vdtor_49A6C0(s32 flags)
     return this;
 }
 
-void ThrowableArray::Remove_49AA00(s16 count)
+void ThrowableArray::Remove_49AA00(s32 count)
 {
     field_20_count -= count;
     if (field_20_count > 0)
@@ -206,7 +206,7 @@ void ThrowableArray::vScreenChange_49AAA0()
     }
 }
 
-void ThrowableArray::Add_49A7A0(s16 count)
+void ThrowableArray::Add_49A7A0(s32 count)
 {
     if (field_6_flags.Get(BaseGameObject::eDead_Bit3))
     {
